Connection: Separate the remote_authorize method from Jetpack and refactor
The remote_authorize method depends on Jetpack. We're currently working on
separating the Connection package from Jetpack, and this commit is part of
that work.

Refactor remote_authorize so that it's separated from Jetpack and fits into
the structure of the Connection package.

* Add the jetpack.remoteAuthorize XMLRPC method to the Connection package's
  XMLRPC_Connector class. Also add a remote_authorize() method to this class.
* Add a handle_authorization() method to the Manager class. This method does
  the work that was previously done by remote_authorize and will be called by
  the XMLRPC remote_authorize methods. Include do_action() calls for tracking
  events.
* Add a convert_to_ixr_error() helper function that converts WP_Error objects to
  IXR_Error objects.
* Add the remote_authorize events to the Tracking class.

NOTE: These changes do not completely separate the handle_authorize method from
Jetpack. The handle_authorize method still uses the Jetpack_Client_Server class,
and that will be addressed in another commit.