Connection: Add a new function which provides the Jetpack API version
Add a new function which provides the Jetpack API version to the Util class. It
will return the JETPACK__API_VERSION constant if it's available. Otherwise, it will
return a default value of 1.

Use this new function to obtain the Jetpack API version in the Manager and Client
classes.

Add Automattic\Jetpack\Connection\Utils to the autoloader ignore list. This is just
temporary to keep the debug log from being filled with PHP notices; it should be
removed when Automattic\Jetpack\Connection\Manager is removed from the ignore list.