Fixes the package release script.
After a recent refactoring a problem has surfaced. The package release script uses a feature of git that allows us to rewrite the tree to only include the needed package at the root. This has worked well before, but now that we have added the package version script, it broke. The reason is that after calling `git filter-branch`, the files get replaced, and the call to `bin/version-packages.sh` fails, because there's no such file anymore.

This change does two things:
* moves the `version-packages.sh` call before we call `git filter-branch`;
* adds a commit call so that the version changes persist and allow calling `git filter-branch`.